# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "inkentry-server",
    "description": "Shared memory server for spelunk. Stores decisions, requirements, and context for a team and serves them over HTTP.",
    "contact": {
      "name": "spelunk",
      "url": "https://github.com/spelunk-cloud/spelunk"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/explore": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Run an LLM reasoning loop over caller-supplied context chunks.\nThe CLI retrieves relevant chunks from its local index and sends them alongside\nthe question. **The server does not store context chunks.**",
        "description": "Returns an SSE stream with events: `thought`, `answer`, `done`, `error`.\nReturns 503 if no LLM is configured.",
        "operationId": "explore",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExploreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream: thought/answer/done/error events"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No LLM configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/projects/{project_id}/llm/complete": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Run a single LLM completion over caller-supplied messages. Streaming SSE.",
        "description": "The server performs no orchestration, adds no system prompt, and stores nothing.\nClient-supplied `max_tokens` is clamped server-side to the configured ceiling.\n\n**Auth:** `Authorization: Bearer` required (Tier 1).\n\n**SSE event shapes:**\n- `{\"kind\":\"token\",\"content\":\"...\"}`: one streamed fragment\n- `{\"kind\":\"done\"}`: terminal success\n- `{\"kind\":\"error\",\"code\":\"...\",\"message\":\"...\"}`: terminal failure mid-stream\n\nReturns 503 if no LLM is configured on this server.",
        "operationId": "llm_complete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream: token/done/error events"
          },
          "400": {
            "description": "messages empty or max_tokens ≤ 0",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No LLM configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Consistent JSON error body: `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ExploreContextChunk": {
        "type": "object",
        "description": "A single context chunk supplied by the CLI for `/explore`.",
        "required": [
          "file",
          "start_line",
          "end_line",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "end_line": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "file": {
            "type": "string"
          },
          "start_line": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ExploreRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/projects/{project_id}/explore`.",
        "required": [
          "question"
        ],
        "properties": {
          "context_chunks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExploreContextChunk"
            }
          },
          "max_turns": {
            "type": "integer",
            "minimum": 0
          },
          "question": {
            "type": "string"
          }
        }
      },
...
      "LlmCompleteMessage": {
        "type": "object",
        "description": "A single chat message for `/llm/complete`.",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "description": "Role: `system`, `user`, or `assistant`."
          }
        }
      },
      "LlmCompleteRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/projects/{project_id}/llm/complete`.",
        "required": [
          "messages",
          "max_tokens"
        ],
        "properties": {
          "json_schema": {
            "description": "Optional OpenAI-style `response_format.json_schema` for structured output."
          },
          "max_tokens": {
            "type": "integer",
            "description": "Desired max completion tokens. The server clamps this to its configured ceiling.",
            "minimum": 0
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LlmCompleteMessage"
            },
            "description": "Non-empty list of chat messages."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Pass as `Authorization: Bearer <key>`. Not required when no key is configured on the server."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}